04- Deep Learning/Fashion_MNIST_Project/requirements.txt
fastapi==0.123.9
uvicorn==0.38.0
pydantic==2.12.5
tensorflow==2.20.0
scikit-learn==1.7.2
joblib==1.4.2
python-dotenv==1.0.1
python-multipart==0.0.20
